+++ placetopaypayment.php
@@ -4,7 +4,7 @@ if (!defined('_PS_VERSION_')) {
     exit;
 }
 
-require_once 'spl_autoload.php';
+require_once __DIR__ . '/vendor/autoload.php';
 
 class PlacetoPayPayment extends PlacetoPay\Models\PlacetoPayPayment
 {