+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - IMPORTANT: Notification URL used in Panel required changes to support new controllers.
 
 ### Removed
-- Drop support for Prestashop <= 1.7.
+- Drop support for Prestashop <= 1.7.x (minimum is now 8.x).
 
 ## [5.0.3 (2026-02-24)](https://github.com/placetopay/prestashop-placetopay/compare/5.0.2...5.0.3)